Make "blank" question type or blank input box use a predefined question.
Currently you can only have 1 blank question due to the fact that the
ID is the same, so it just updates the value of it (number what the
question is). Swapping the key within the question list for the quiz to
the number of the question and put the question ID as the value instead of
the number of question may fix the issue since the list is unique
 (1-infinity) to the specific quiz so there should be no overlap in
numbers so they can be used as keys.