Editing a question has been implemented.
The question is edited by pressing the "Breyta" button next to the
question that is to be edited. That'll make the list element editable
and when the button is pressed again (now labeled "Vista") it'll check
whether the question is private or not. If it is private then the
question is simply updated with the new value. On the other hand if
the question is not private, a duplicate of the question is created
which is private and replaces the public question in the quiz with the
newly changed now private question.

There are 2 issues with this code.
 - The entire list element becomes editable so the buttons that come
   with said list element also become editable which is not desired.
 - Due to issue 1, if the user clears out the entire question he starts
   editing the button and can't seem to go back to editing the now
   empty question (refreshing brings back the old question since
   the code won't allow for empty questions to be the "update").

If this for some reason is a desired result (for the buttons to be
editable) then the regex has to be changed to reflect possible
changes to the button text.